Open metadata restrict for user role metadata admin only

+++ Kartverket.MetadataEditor/Controllers/OpenMetadataEndpointsController.cs
@@ -24,12 +24,18 @@ namespace Kartverket.MetadataEditor.Controllers
         // GET: OpenMetadataEndpoints
         public ActionResult Index()
         {
+            if(!IsAdmin())
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.Forbidden);
+
             return View(_db.OpenMetadataEndpoints.ToList());
         }
 
         // GET: OpenMetadataEndpoints/Details/5
         public ActionResult Details(int? id)
         {
+            if (!IsAdmin())
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.Forbidden);
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -45,6 +51,9 @@ namespace Kartverket.MetadataEditor.Controllers
         // GET: OpenMetadataEndpoints/Create
         public ActionResult Create()
         {
+            if (!IsAdmin())
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.Forbidden);
+
             return View();
         }
 
@@ -55,6 +64,9 @@ namespace Kartverket.MetadataEditor.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Url,OrganizationName")] OpenMetadataEndpoint openMetadataEndpoint)
         {
+            if (!IsAdmin())
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.Forbidden);
+
             if (ModelState.IsValid)
             {
                 _db.OpenMetadataEndpoints.Add(openMetadataEndpoint);
@@ -68,6 +80,9 @@ namespace Kartverket.MetadataEditor.Controllers
         // GET: OpenMetadataEndpoints/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (!IsAdmin())
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.Forbidden);
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -87,6 +102,9 @@ namespace Kartverket.MetadataEditor.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Url,OrganizationName")] OpenMetadataEndpoint openMetadataEndpoint)
         {
+            if (!IsAdmin())
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.Forbidden);
+
             if (ModelState.IsValid)
             {
                 _db.Entry(openMetadataEndpoint).State = EntityState.Modified;
@@ -99,6 +117,9 @@ namespace Kartverket.MetadataEditor.Controllers
         // GET: OpenMetadataEndpoints/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (!IsAdmin())
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.Forbidden);
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -116,10 +137,54 @@ namespace Kartverket.MetadataEditor.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!IsAdmin())
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.Forbidden);
+
             OpenMetadataEndpoint openMetadataEndpoint = _db.OpenMetadataEndpoints.Find(id);
             _db.OpenMetadataEndpoints.Remove(openMetadataEndpoint);
             _db.SaveChanges();
             return RedirectToAction("Index");
+        }
+
+        bool IsAdmin()
+        {
+            if (User.Identity.IsAuthenticated)
+            {
+                string userOrganization = GetSecurityClaim("organization");
+                string role = GetSecurityClaim("role");
+                if (!string.IsNullOrWhiteSpace(role) && role.Equals("nd.metadata_admin"))
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
+        private string GetUsername()
+        {
+            return GetSecurityClaim("username");
+        }
+
+        private string GetSecurityClaim(string type)
+        {
+            string result = null;
+            foreach (var claim in System.Security.Claims.ClaimsPrincipal.Current.Claims)
+            {
+                if (claim.Type == type && !string.IsNullOrWhiteSpace(claim.Value))
+                {
+                    result = claim.Value;
+                    break;
+                }
+            }
+
+            // bad hack, must fix BAAT
+            if (!string.IsNullOrWhiteSpace(result) && type.Equals("organization") && result.Equals("Statens kartverk"))
+            {
+                result = "Kartverket";
+            }
+
+            return result;
         }
 
         protected override void Dispose(bool disposing)
